Add selected project Svelte store
Introduces a Svelte store to manage the selected project state, modal open/close status, and focus restoration. This enables tracking and updating the currently selected project and improves accessibility by restoring focus after modal closure.

NOTE: If I wasn't under a tight timeline to do this quickly, I'd probably reconsider having the selected projects done in a store in this way. I might just make it reactive, but for now, this is fine. It might have even been extra work to do it this way. 🤷‍♂️ In theory, I could start integrating the navigation buttons forward/backward to go back and forth between state, and this might help me do that. Anyway, that was my thinking...